+++ nucleus/dataset.py
@@ -43,6 +43,7 @@ from .constants import (
     DATASET_IS_SCENE_KEY,
     DATASET_ITEM_IDS_KEY,
     DATASET_ITEMS_KEY,
+    DATASET_PRIVACY_MODE_KEY,
     DEFAULT_ANNOTATION_UPDATE_MODE,
     EMBEDDING_DIMENSION_KEY,
     EMBEDDINGS_URL_KEY,
@@ -75,6 +76,7 @@ from .dataset_item import (
     DatasetItem,
     check_all_paths_remote,
     check_for_duplicate_reference_ids,
+    check_items_have_dimensions,
 )
 from .dataset_item_uploader import DatasetItemUploader
 from .deprecation_warning import deprecated
@@ -145,12 +147,20 @@ class Dataset:
         existing_dataset = client.get_dataset("YOUR_DATASET_ID")
     """
 
-    def __init__(self, dataset_id, client: "NucleusClient", name=None):
+    def __init__(
+        self,
+        dataset_id,
+        client: "NucleusClient",
+        name=None,
+        is_scene=None,
+        use_privacy_mode=None,
+    ):
         self.id = dataset_id
         self._client = client
         # NOTE: Optionally set name on creation such that the property access doesn't need to hit the server
         self._name = name
-        self._is_scene = None
+        self._is_scene = is_scene
+        self._use_privacy_mode = use_privacy_mode
 
     def __repr__(self):
         if os.environ.get("NUCLEUS_DEBUG", None):
@@ -183,6 +193,17 @@ class Dataset:
         )[DATASET_IS_SCENE_KEY]
         self._is_scene = response
         return self._is_scene  # type: ignore
+
+    @property
+    def use_privacy_mode(self) -> bool:
+        """Whether or not the dataset was created for privacy mode."""
+        if self._use_privacy_mode is not None:
+            return self._use_privacy_mode
+        response = self._client.make_request(
+            {}, f"dataset/{self.id}/use_privacy_mode", requests.get
+        )[DATASET_PRIVACY_MODE_KEY]
+        self._use_privacy_mode = response
+        return self._use_privacy_mode  # type: ignore
 
     @property
     def model_runs(self) -> List[str]:
@@ -655,6 +676,9 @@ class Dataset:
         video_scenes = [item for item in items if isinstance(item, VideoScene)]
 
         check_for_duplicate_reference_ids(dataset_items)
+
+        if self.use_privacy_mode:
+            check_items_have_dimensions(dataset_items)
 
         if dataset_items and (lidar_scenes or video_scenes):
             raise Exception(